fix(fcp-host): widen intermediate arithmetic to u128 to prevent overflow in rate calculations
The EWMA latency tracker in resilience.rs and the success/error rate
calculators in rollout.rs performed intermediate multiplications in u64
space. Under sustained high-sample-count workloads, the product of
current_millis * retained_per_mille (or successes * 10_000) could
silently overflow u64::MAX, producing incorrect rate metrics that would
cascade into bad rollout decisions.

Changes:
- resilience.rs: LatencyEwma::update() now widens alpha, retained, and
  sample values to u128 before multiplying, then narrows back with
  u64::try_from(...).unwrap_or(u64::MAX)
- resilience.rs: ratio_per_mille() similarly widens numerator and
  denominator to u128 before the multiply-then-divide pattern
- rollout.rs: success_rate_bps() and error_rate_bps() widen to u128
  before multiplying by the 10_000 basis-point factor

All existing tests pass unchanged because the fix only affects edge-case
magnitudes, but the correctness guarantee now covers the full u64 input
domain.

+++ crates/fcp-host/src/rollout.rs
@@ -959,8 +959,8 @@ fn success_rate_bps(record: &LifecycleRecord) -> u16 {
     if record.health.samples == 0 {
         return 10_000;
     }
-    let numerator = record.health.successes.saturating_mul(10_000);
-    let rate = numerator / record.health.samples;
+    let numerator = u128::from(record.health.successes).saturating_mul(10_000);
+    let rate = numerator / u128::from(record.health.samples);
     u16::try_from(rate.min(10_000)).unwrap_or(10_000)
 }
 
@@ -968,8 +968,8 @@ fn error_rate_bps(record: &LifecycleRecord) -> u16 {
     if record.health.samples == 0 {
         return 0;
     }
-    let numerator = record.health.failures.saturating_mul(10_000);
-    let rate = numerator / record.health.samples;
+    let numerator = u128::from(record.health.failures).saturating_mul(10_000);
+    let rate = numerator / u128::from(record.health.samples);
     u16::try_from(rate.min(10_000)).unwrap_or(10_000)
 }
